Add support for condition function in set value API (#78)
This PR adds support for supplying a "condition" function callback to the hash table set value API which allows the caller to disallow the set value operation based on some caller-defined criteria. The callback is provided access to the existing key and the new key being provided via the callback. This might seem counter-intuitive, but consider cases where the "key" is a composite structure where the equivalence is established by a subset of the fields but other fields come into play in other circumstances like when an existing value is being replaced.